Move benchmark sub-packge to new top-level benchmarks directory
