+++ .circleci/config.yml
@@ -107,7 +107,7 @@ jobs:
           name: run tests benchmark
           command: |
             . venv/bin/activate
-            python -m xain.benchmark.exec --group_name=ci-test-group --task_name=ci-test-task --dataset=fashion-mnist-100p-iid-balanced --model=blog_cnn --R=2 --E=2 --C=0.02 --B=64
+            python -m benchmarks.benchmark.exec --group_name=ci-test-group --task_name=ci-test-task --dataset=fashion-mnist-100p-iid-balanced --model=blog_cnn --R=2 --E=2 --C=0.02 --B=64
 
   # Test if docker image can be build without failing
   docker_build:
