+++ benchmarks/benchmark/aggregation/plot.py
