+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-testpaths = xain
+testpaths = benchmarks xain
 markers =
     unmarked: every test without a marker will automatically have this marker assigned. You could senselessly add it manually
     integration: mark a integration test
